# API文档/default_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "瑞吉外卖接口文档",
    "version": "1.0",
    "title": "瑞吉外卖"
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
    "/addressBook": {
      "post": {
        "tags": [
          "address-book-controller"
        ],
        "summary": "save",
        "operationId": "saveUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "addressBook",
            "description": "addressBook",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddressBook",
              "originalRef": "AddressBook"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«AddressBook»",
              "originalRef": "R«AddressBook»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "address-book-controller"
        ],
        "summary": "update",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "addressBook",
            "description": "addressBook",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddressBook",
              "originalRef": "AddressBook"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»",
              "originalRef": "R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/addressBook/default": {
      "get": {
        "tags": [
          "address-book-controller"
        ],
        "summary": "getDefault",
        "operationId": "getDefaultUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«AddressBook»",
              "originalRef": "R«AddressBook»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "address-book-controller"
        ],
        "summary": "setDefault",
        "operationId": "setDefaultUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "addressBook",
            "description": "addressBook",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddressBook",
              "originalRef": "AddressBook"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«AddressBook»",
              "originalRef": "R«AddressBook»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/addressBook/list": {
      "get": {
        "tags": [
          "address-book-controller"
        ],
        "summary": "list",
        "operationId": "listUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "cityCode",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "cityName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "consignee",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "createTime",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "createUser",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "detail",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "districtCode",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "districtName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "isDefault",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "label",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "provinceCode",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "provinceName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sex",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "updateTime",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updateUser",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«List«AddressBook»»",
              "originalRef": "R«List«AddressBook»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/addressBook/{id}": {
      "get": {
        "tags": [
          "address-book-controller"
        ],
        "summary": "get",
        "operationId": "getUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«AddressBook»",
              "originalRef": "R«AddressBook»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddressBook": {
      "type": "object",
      "properties": {
        "cityCode": {
          "type": "string",
          "refType": null
        },
        "cityName": {
          "type": "string",
          "refType": null
        },
        "consignee": {
          "type": "string",
          "refType": null
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "detail": {
          "type": "string",
          "refType": null
        },
        "districtCode": {
          "type": "string",
          "refType": null
        },
        "districtName": {
          "type": "string",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "isDefault": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "label": {
          "type": "string",
          "refType": null
        },
        "phone": {
          "type": "string",
          "refType": null
        },
        "provinceCode": {
          "type": "string",
          "refType": null
        },
        "provinceName": {
          "type": "string",
          "refType": null
        },
        "sex": {
          "type": "string",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "userId": {
          "type": "integer",
          "format": "int64",
          "refType": null
        }
      },
      "title": "AddressBook"
    },
...
    "R«AddressBook»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/AddressBook",
          "originalRef": "AddressBook"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«AddressBook»"
    },
...
    "R«List«AddressBook»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AddressBook",
            "originalRef": "AddressBook"
          }
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«List«AddressBook»»"
    },
...
    "R«string»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "string"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«string»"
    },
...
  }
}